+++ app/src/main/res/layout/fragment_3.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/noah_6"
-    tools:context="Fragment_3"
+    tools:context=".activities.BinaryCalculatorFragment"
     tools:ignore="HardcodedText"
     tools:layout_editor_absoluteY="81dp">
     <!--android:background="@color/dark_grey_pressed_mask" -->
